+++ Sources/SwiftExtras/Extensions/Image+Platform.swift
@@ -95,4 +95,3 @@ extension PlatformImage: Identifiable {
 }
 #endif
 #endif
-